+++ Services/ServiceRol.cs
@@ -3,25 +3,23 @@ using API_MortalKombat.Repository.IRepository;
 using API_MortalKombat.Services.IService;
 using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
-using System.Net;
 using API_MortalKombat.Models.DTOs.RolDTO;
 using Microsoft.AspNetCore.JsonPatch;
-using API_MortalKombat.Models.DTOs.ReinoDTO;
 using FluentValidation;
 using API_MortalKombat.Services.Utils;
 
 namespace API_MortalKombat.Service
 {
-    public class ServiceRol : IServiceRol
+    public class ServiceRol : IServiceGeneric<RolUpdateDto, RolCreateDto>
     {
-        private readonly IRepositoryRol _repository;
-        private readonly IRepositoryUsuario _repositoryUsuario;
+        private readonly IRepositoryGeneric<Rol> _repository;
+        private readonly IRepositoryGeneric<Usuario> _repositoryUsuario;
         private readonly IMapper _mapper;
         private readonly APIResponse _apiresponse;
         private readonly ILogger<ServiceRol> _logger;
         private readonly IValidator<RolCreateDto> _validator;
         private readonly IValidator<RolUpdateDto> _validatorUpdate;
-        public ServiceRol(IMapper mapper, APIResponse apiresponse, ILogger<ServiceRol> logger, IRepositoryRol repository, IRepositoryUsuario repositoryUsuario, IValidator<RolCreateDto> 
+        public ServiceRol(IMapper mapper, APIResponse apiresponse, ILogger<ServiceRol> logger, IRepositoryGeneric<Rol> repository, IRepositoryGeneric<Usuario> repositoryUsuario, IValidator<RolCreateDto> 
                           validator, IValidator<RolUpdateDto> validatorUpdate)
         {
             _mapper = mapper;
@@ -33,14 +31,12 @@ namespace API_MortalKombat.Service
             _validatorUpdate = validatorUpdate;
         }
 
-        public async Task<APIResponse> GetRoles()
+        public async Task<APIResponse> GetAll()
         {
             try
             {
                 List<Rol> listRoles = await _repository.GetAll();
-                _apiresponse.Result = _mapper.Map<IEnumerable<RolDto>>(listRoles);
-                _apiresponse.statusCode = HttpStatusCode.OK;
-                return _apiresponse;
+                Utils.ListCorrectResponse<RolDto, Rol>(_mapper, listRoles, _apiresponse);
             }
             catch (Exception ex)
             {
@@ -49,18 +45,16 @@ namespace API_MortalKombat.Service
             return _apiresponse;
         }
 
-        public async Task<APIResponse> GetRolById(int id)
+        public async Task<APIResponse> GetById(int id)
         {
             try
             {
                 var rol = await _repository.GetById(id);
-                if (Utils.CheckIfNull(rol, _apiresponse, _logger) != null)
+                if (!Utils.CheckIfNull(rol, _apiresponse, _logger))
                 {
                     return _apiresponse;
                 }
-                _apiresponse.Result = _mapper.Map<RolDto>(rol);
-                _apiresponse.statusCode = HttpStatusCode.OK;
-                return _apiresponse;
+                Utils.CorrectResponse<RolDto, Rol>(_mapper, rol, _apiresponse);
             }
             catch (Exception ex)
             {
@@ -69,18 +63,16 @@ namespace API_MortalKombat.Service
             return _apiresponse;
         }
 
-        public async Task<APIResponse> GetRolByName(string name)
+        public async Task<APIResponse> GetByName(string name)
         {
             try
             {
                 var rol = await _repository.GetByName(name);
-                if (Utils.CheckIfNull(rol, _apiresponse, _logger) != null)
+                if (!Utils.CheckIfNull(rol, _apiresponse, _logger))
                 {
                     return _apiresponse;
                 }
-                _apiresponse.Result = _mapper.Map<RolDto>(rol);
-                _apiresponse.statusCode = HttpStatusCode.OK;
-                return _apiresponse;
+                Utils.CorrectResponse<RolDto, Rol>(_mapper, rol, _apiresponse);
             }
             catch (Exception ex)
             {
@@ -89,7 +81,7 @@ namespace API_MortalKombat.Service
             return _apiresponse;
         }
 
-        public async Task<APIResponse> CreateRol([FromBody] RolCreateDto rolCreateDto)
+        public async Task<APIResponse> Create([FromBody] RolCreateDto rolCreateDto)
         {
             try
             {
@@ -97,20 +89,16 @@ namespace API_MortalKombat.Service
                 {
                     return _apiresponse;
                 }
-                var existeRol = await _repository.GetByName(rolCreateDto.Nombre);
-                if (existeRol != null)
+                var existRol = await _repository.GetByName(rolCreateDto.Nombre);
+                if (!Utils.CheckIfObjectExist<Rol>(existRol, _apiresponse, _logger))
                 {
-                    _apiresponse.isExit = false;
-                    _apiresponse.statusCode = HttpStatusCode.Conflict;
-                    _logger.LogError("Ya existe un rol con el mismo nombre.");
                     return _apiresponse;
                 }
                 var rol = _mapper.Map<Rol>(rolCreateDto);
                 rol!.FechaCreacion = DateTime.Now;
                 await _repository.Create(rol);
-                _apiresponse.statusCode = HttpStatusCode.OK;
-                _apiresponse.Result = _mapper.Map<RolDto>(rol);
                 _logger.LogInformation("¡Rol creado con exito!");
+                Utils.CorrectResponse<RolDto, Rol>(_mapper, rol, _apiresponse);
             }
             catch (Exception ex)
             {
@@ -119,31 +107,24 @@ namespace API_MortalKombat.Service
             return _apiresponse;
         }
 
-        public async Task<APIResponse> DeleteRol(int id)
+        public async Task<APIResponse> Delete(int id)
         {
             try
             {
                 var rol = await _repository.GetById(id);
-                if (Utils.CheckIfNull(rol, _apiresponse, _logger) != null)
+                if (!Utils.CheckIfNull(rol, _apiresponse, _logger))
                 {
                     return _apiresponse;
                 }
                 var listUsuarios = await _repositoryUsuario.GetAll();
-                foreach (var i in listUsuarios) //aqui podria usarse el metodo cascada para que se borre todo, pero decidi agergarle esto para mas seguridad
+                if (!Utils.PreventDeletionIfRelaredUserExist(listUsuarios, _apiresponse, id))
                 {
-                    if (i.RolId == id)
-                    {
-                        _apiresponse.statusCode = HttpStatusCode.NotFound;
-                        _logger.LogError("El Rol no se puede eliminar porque el usuario " + i.NombreDeUsuario + " de id " + i.Id + " contiene como RolId este rol.");
-                        _apiresponse.isExit = false;
-                        return _apiresponse;
-                    }
+                    _logger.LogError("El rol no se puede eliminar porque hay un usuario que contiene como RolId este rol.");
+                    return _apiresponse;
                 }
                 await _repository.Delete(rol);
-                _apiresponse.statusCode = HttpStatusCode.OK;
-                _apiresponse.Result = _mapper.Map<RolDto>(rol);
                 _logger.LogInformation("!Rol eliminado con exito¡");
-                return _apiresponse;
+                Utils.CorrectResponse<RolDto, Rol>(_mapper, rol, _apiresponse);
             }
             catch (Exception ex)
             {
@@ -152,7 +133,7 @@ namespace API_MortalKombat.Service
             return _apiresponse;
         }
 
-        public async Task<APIResponse> UpdateRol(int id, [FromBody] RolUpdateDto rolUpdateDto)
+        public async Task<APIResponse> Update([FromBody] RolUpdateDto rolUpdateDto)
         {
             try
             {
@@ -160,36 +141,21 @@ namespace API_MortalKombat.Service
                 {
                     return _apiresponse;
                 }
-                if (id == 0 || id != rolUpdateDto.Id)
+                var rol = await _repository.GetById(rolUpdateDto.Id);
+                if (!Utils.CheckIfNull<Rol>(rol, _apiresponse, _logger))
                 {
-                    _apiresponse.isExit = false;
-                    _apiresponse.statusCode = HttpStatusCode.NotFound;
-                    _logger.LogError("Error con la id ingresada.");
-                    return _apiresponse;
-                }
-                var existRol = await _repository.GetById(id);
-                if (existRol == null)
-                {
-                    _apiresponse.isExit = false;
-                    _apiresponse.statusCode = HttpStatusCode.NotFound;
-                    _logger.LogError("No se encuentra registrado el id ingresado.");
                     return _apiresponse;
                 }
                 var registredName = await _repository.GetByName(rolUpdateDto.Nombre);
-                if (registredName != null && registredName.Id != rolUpdateDto.Id)
+                if (!Utils.CheckIfNameAlreadyExist<Rol>(registredName, rol, _apiresponse, _logger))
                 {
-                    _apiresponse.isExit = false;
-                    _apiresponse.statusCode = HttpStatusCode.Conflict;
-                    _logger.LogError("Ya existe un rol con el mismo nombre.");
                     return _apiresponse;
                 }
-                _mapper.Map(rolUpdateDto, existRol);
-                existRol.FechaActualizacion = DateTime.Now;
-                _apiresponse.statusCode = HttpStatusCode.OK;
-                _apiresponse.Result = _mapper.Map<RolDto>(existRol);
+                _mapper.Map(rolUpdateDto, rol);
+                rol.FechaActualizacion = DateTime.Now;
+                await _repository.Update(rol);
                 _logger.LogInformation("¡Rol Actualizado con exito!");
-                await _repository.Update(existRol);
-                return _apiresponse;
+                Utils.CorrectResponse<RolDto, Rol>(_mapper, rol, _apiresponse);                
             }
             catch (Exception ex)
             {
@@ -198,12 +164,12 @@ namespace API_MortalKombat.Service
             return _apiresponse;
         }
 
-        public async Task<APIResponse> UpdatePartialRol(int id, JsonPatchDocument<RolUpdateDto> rolUpdateDto)
+        public async Task<APIResponse> UpdatePartial(int id, JsonPatchDocument<RolUpdateDto> rolUpdateDto)
         {
             try
             {
                 var rol = await _repository.GetById(id);
-                if (Utils.CheckIfNull(rol, _apiresponse, _logger) != null)
+                if (!Utils.CheckIfNull(rol, _apiresponse, _logger))
                 {
                     return _apiresponse;
                 }
@@ -214,19 +180,15 @@ namespace API_MortalKombat.Service
                     return _apiresponse;
                 }
                 var registredName = await _repository.GetByName(updateRolDto.Nombre); //verifico que no haya otro con el mismo nomrbe
-                if (registredName != null && registredName.Id != rol.Id)
+                if (!Utils.CheckIfNameAlreadyExist<Rol>(registredName, rol, _apiresponse, _logger))
                 {
-                    _apiresponse.isExit = false;
-                    _apiresponse.statusCode = HttpStatusCode.Conflict; // Conflict indica que ya existe un recurso con el mismo nombre
-                    _logger.LogError("Ya existe un rol con el mismo nombre.");
                     return _apiresponse;
                 }
                 _mapper.Map(updateRolDto, rol);
                 rol.FechaActualizacion = DateTime.Now;
                 await _repository.Update(rol);
-                _apiresponse.statusCode = HttpStatusCode.OK;
-                _apiresponse.Result = updateRolDto;
-                return _apiresponse;
+                _logger.LogInformation("¡Rol Actualizado con exito!");
+                Utils.CorrectResponse<RolDto, Rol>(_mapper, rol, _apiresponse);
             }
             catch (Exception ex)
             {
